Rename NotMe to NotSender, add pluginConfig to AccountPaymentService
