+++ eternaleconomy-core/src/main/java/com/eternalcode/economy/command/validator/notsender/NotSenderValidator.java
@@ -1,4 +1,4 @@
-package com.eternalcode.economy.command.validator.notyourself;
+package com.eternalcode.economy.command.validator.notsender;
 
 import com.eternalcode.economy.account.Account;
 import com.eternalcode.economy.config.implementation.messages.MessageConfig;
@@ -10,18 +10,18 @@ import dev.rollczi.litecommands.validator.ValidatorResult;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 
-public class NotMeValidator implements AnnotatedValidator<CommandSender, Account, NotMe> {
+public class NotSenderValidator implements AnnotatedValidator<CommandSender, Account, NotSender> {
 
     private final MessageConfig config;
 
-    public NotMeValidator(MessageConfig config) {
+    public NotSenderValidator(MessageConfig config) {
         this.config = config;
     }
 
     @Override
-    public ValidatorResult validate(Invocation<CommandSender> invocation, CommandExecutor<CommandSender> commandExecutor, Requirement<Account> requirement, Account account, NotMe notMe) {
+    public ValidatorResult validate(Invocation<CommandSender> invocation, CommandExecutor<CommandSender> commandExecutor, Requirement<Account> requirement, Account account, NotSender notSender) {
         if (invocation.sender() instanceof Player sender && sender.getUniqueId().equals(account.uuid())) {
-            return ValidatorResult.invalid(config.notMe);
+            return ValidatorResult.invalid(this.config.notSender);
         }
 
         return ValidatorResult.valid();
